requirements.txt
opencv-python
mediapipe
numpy
pyautogui
PyQt5
pynput
edge-tts
googlesearch-python
requests
psutil
SpeechRecognition
pyaudio
pygetwindow
pygame
pyttsx3
pyperclip
ollama
groq
Pillow
python-dotenv
dateparser
faster-whisper
torch
torchaudio
kokoro-onnx
soundfile
beautifulsoup4
feedparser
pywin32

keyboard
